// frontend/src/app/generated/api/model/applicationReminderViewModel.ts
/**
 * Yaam.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApplicationReminderViewModelDelayDays } from './applicationReminderViewModelDelayDays';

export interface ApplicationReminderViewModel {
  id: string;
  delayDays: ApplicationReminderViewModelDelayDays;
  dueDate: string;
  note: string | null;
  notifiedAt: string | null;
  completedAt: string | null;
  createdAt: string;
}